pyproject.toml
[tool.poetry]
name = "kmua"
version = "2.1.0"
description = ""
authors = ["krau <me@krau.top>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = { extras = ["rate-limiter"], version = "^20.6" }
dynaconf = "^3.2.3"
graphviz = "^0.20.1"
Pillow = "9.5.0"
anyio = "^4.0.0"
apscheduler = "^3.10.4"
certifi = "^2023.7.22"
pilmoji = "^2.0.4"
pypinyin = "^0.49.0"
pytz = "^2023.3.post1"
six = "^1.16.0"
sqlalchemy = "^2.0.21"
httpx = "^0.25.0"
zhconv = "^1.4.3"
psutil = "^5.9.5"
tzlocal = "^5.1"
typing-extensions = "^4.8.0"
sniffio = "^1.3.0"
idna = "^3.4"
loguru = "^0.7.2"
greenlet = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
